DEV-1925: Checkout app - decouple card creation and payout generation
